% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_package_table.R
\name{generate_package_table}
\alias{generate_package_table}
\title{Generate a table with R package information}
\usage{
generate_package_table(repo_paths)
}
\arguments{
\item{repo_paths}{A character vector with GH paths to the R package
repositories}
}
\value{
A tibble with the following columns:
\itemize{
\item Repo
\item Lifecycle
\item Status
\item Latest_SHA
\item Maintainer
}
}
\description{
Generate a table with R package information
}
\examples{
generate_package_table(c("RMI-PACTA/r2dii.data", "RMI-PACTA/r2dii.match"))
}
