% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_workflow_table.R
\name{generate_workflow_table}
\alias{generate_workflow_table}
\title{Generate a table with worfklow information}
\usage{
generate_workflow_table(repo_paths)
}
\arguments{
\item{repo_paths}{A character vector with GH paths to the workflow
repositories}
}
\value{
A tibble with the following columns:
\itemize{
\item Repo
\item Lifecycle
\item Docker Status
\item Latest_SHA
\item Maintainer
}
}
\description{
Generate a table with worfklow information
}
\examples{
generate_workflow_table("RMI-PACTA/workflow.transition.monitor")
}
