% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_badge_tables.R
\name{generate_package_table}
\alias{generate_package_table}
\title{Generate a table with R package information}
\usage{
generate_package_table(repo_vectors)
}
\arguments{
\item{repo_vectors}{A list of named character vectors. Each character vector
should contain:
\itemize{
\item path - the shortpath to the GitHub repository, incl. org (e.g. "RMI-PACTA/r2dii.data")
\item lifecycle - the lifecycle badge status (e.g. "stable", "experimental")
\item ci_check - the name of the CI check file (e.g. "R-CMD-check.yaml")
}}
}
\value{
A formatted sitrep tibble.
}
\description{
Generate a table with R package information
}
\examples{

repos <- list(
  c(
    path = "RMI-PACTA/r2dii.data",
    lifecycle = "stable",
    ci_check = "R-CMD-check.yaml"
  ),
  c(
    path = "RMI-PACTA/r2dii.plot",
    lifecycle = "stable",
    ci_check = "R.yml"
  )
)

generate_package_table(repos)
}
