% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_badge_tables.R
\name{generate_workflow_table}
\alias{generate_workflow_table}
\title{Generate a table with workflow information}
\usage{
generate_workflow_table(repo_vectors)
}
\arguments{
\item{repo_vectors}{A list of named character vectors. Each character vector
should contain:
\itemize{
\item path - the shortpath to the GitHub repository, including organization
(e.g. "RMI-PACTA/workflow.transition.monitor")
\item lifecycle - the lifecycle badge status (e.g. "stable", "experimental")
\item ci_check - the name of the CI check file
(e.g. "build-Docker-image-triggers.yml")
}}
}
\value{
A formatted sitrep tibble.
}
\description{
Generate a table with workflow information
}
\examples{

repos <- list(
  c(
    path = "RMI-PACTA/workflow.transition.monitor",
    lifecycle = "stable",
    ci_check = "build-Docker-image-triggers.yml"
  ),
  c(
    path = "RMI-PACTA/workflow.data.preparation",
    lifecycle = "stable",
    ci_check = "docker.yml"
  )
)

generate_workflow_table(repos)
}
